backend/requirements.txt
fastapi
uvicorn
sqlmodel
alembic
pymysql
python-dotenv
pydantic-settings
pytest
httpx
requests
markdown
pdfkit
PyJWT
python-multipart